Hack out unicode_util_compat
unicode_util_compat is a polyfill for OTP20's unicode_util.  Trying
to build a release with Distillery on OTP20 with unicode_util_compat
as well causes a conflict.

Don't have time to come up with a proper fix right now, but this
should get us going as long as the library is only used with OTP20+